chore: cleanup playground files
